+++ module/data/item/missileweapon.js
@@ -26,9 +26,14 @@ export class MissileWeaponItemData extends BaseWeaponItemData {
    * If the item has a gunner, then return the gunner's name and skillValue
    */
   get gunnerSkill() {
-    if (this.gunner)
-      return this.gunner.system.skills[this.attackWith];
-    else
-      return { value: 0, adds: 0 }
+    if (this.gunner) {
+      const skill = this.gunner.system.skills[this.attackWith];
+      if (skill) {
+        const result = { ...skill };
+        result.value -= this.gunner.system.wounds.value;
+        return result;
+      }
+    }
+    return { value: 0, adds: 0 }
   }
 }