+++ Changelog.md
@@ -7,6 +7,7 @@
 - #478: Add optional **Piety score** in the XP block of Stormknights (can be disabled in Game Settings). Update the "Buff" macro to support adding/removing Piety from the currently selected Actors.
 - Add **Supernatural Evil** trait which can be set on abilities such as "Undead" (no automated effect yet).
 - Change **Custom Skills** to have the Total be calculated from the ADDS on the skill + the chosen attribute on the owning Actor. Custom skill values (not the name) can be edited directly in the skill list on Stormknight sheets.
+- A vehicle's operator/gunner skill is now modified by the wounds on the assigned actor.
 
 ### Bug Fixes